Fix source bugs found while getting a clean build
None of these are platform-specific; they just hadn't been exercised
by a compiler/toolchain combination that caught them:

- panelanalysis.cpp: several "#elif" preprocessor directives with no
  condition (should have been "#else"), plus a stray trailing comma
  in one of the affected calls, in the LAPACK_FORTRAN_STRLEN_END
  fallback branches.
- testmatrix.cpp: the same fallback branch mistakenly kept the
  trailing string-length argument that only the LAPACK_FORTRAN_-
  STRLEN_END branch is supposed to pass.
- fileio.cpp, utils-io.cpp: QDataStream/QTextStream used without
  including their headers, relying on them being pulled in
  transitively by something else.
- flow5-io.cpp, cadexportdlg.cpp: StdFail_NotDone (and, in flow5-io.cpp,
  Standard_NoSuchObject) caught by name in a try/catch without
  including their OCCT headers.
- popp3dctrls.cpp: QAction used but its #include was commented out.
- stlreaderdlg.cpp: appended EOLstr (a std::string) to a QString via
  operator+=, which QString does not support; the surrounding code
  uses EOLch (the QString equivalent) for exactly this purpose.
- gl3dxplaneview.cpp: std::sinf/std::cosf are not part of the C++
  standard library; use the std::sin/std::cos overloads instead.
- xplane.cpp: declared a std::vector<QPair<...>> where the callee
  expects std::vector<std::pair<...>>; QPair and std::pair are
  distinct types.
- linestyle.h: LineStyle is stored in a QVariant (QVariant::fromValue/
  value<LineStyle>()) in several places but was never registered with
  Q_DECLARE_METATYPE, which only surfaces as a hard compile error at
  the first use site.
- gmesh_globals.cpp: getNumberOption()/getStringOption() let
  gmsh::option::getNumber()/getString() propagate an exception for an
  option that doesn't exist in the installed gmsh version, crashing
  the whole process instead of reporting that one option as
  unavailable.

+++ flow5-app/modules/xplane/glview/gl3dxplaneview.cpp
@@ -2176,7 +2176,7 @@ void gl3dXPlaneView::glMake3dObjects()
         //    a roll gives the bank angle.
         m_matPOpp.setToIdentity();
         m_matPOpp.rotate(pPOpp->beta(),  {0,0,1});
-        m_matPOpp.rotate(pPOpp->alpha(), {-std::sinf(pPOpp->beta()*PI/180.0), std::cosf(pPOpp->beta()*PI/180.0), 0.0f});
+        m_matPOpp.rotate(pPOpp->alpha(), {-std::sin(pPOpp->beta()*PI/180.0), std::cos(pPOpp->beta()*PI/180.0), 0.0f});
     }
 
     if(s_bResetglGeom)
